moving query code to Utilities
